chore(rxSelectFilter): decompose monolithic source
