+++ src/rxSelectFilter/rxSelectFilter.module.js
@@ -1,0 +1,22 @@
+/**
+ * @ngdoc overview
+ * @name rxSelectFilter
+ * @description
+ * # rxSelectFilter component
+ *
+ * A component that provides a multi-select dropdown interface intended for
+ * table filtering.
+ *
+ * ## Directives
+ * * {@link rxSelectFilter.directive:rxSelectFilter rxSelectFilter}
+ *
+ * ## Filters
+ * * {@link rxSelectFilter.filter:Apply Apply}
+ *
+ * ## Services
+ * * {@link rxSelectFilter.service:SelectFilter SelectFilter}
+ */
+angular.module('encore.ui.rxSelectFilter', [
+    'encore.ui.rxMisc',
+    'encore.ui.rxSelect'
+]);
